chatページのSSR化
